Fix Google Chrome Manifest v3 compatibility
The built background.js script contained
few left over instances of window global,
which is not declared in service workers.

On Chrome and other Chromium based browsers
that end up using the service worker variation,
will fail on a fatal error due to the missing
global.

This PR changes so that the combined service worker
or background script checks whether the debugging flag
exits either in the self or the window global in a
graceful manner, accounting for either or being
undeclared.

Also adds missing `scripting` permission to the
manifest. In Chrome, the used
`chrome.scripting.executeScript` function
is only available when the extension
has scripting permissions. Without it,
filling form fields will not function.